+++ test/convertor.test.js
@@ -41,11 +41,17 @@ let run = function () {
     // lower_case_with_underscores
     console.log('通过日期获取星座', Convertor.getConstellationByDate('09/14'));
     console.log('通过日期获取星座', Convertor.getConstellationByDate(new Date(1997, 2, 14)));
+
+    console.log('通过日期获取星座', Convertor.getConstellationByDate(new Date(1997, 2, 14)));
+
+    console.log('身份证号解析', '230504199607116664'.citizenIdentificationNumberParse);
+
 }
 try {
     describe('转换模块', function () {
         it('社会统一信用代码转换组织机构代码', function () {
             expect(Convertor.usciToOibc('91411100766237140X')).toEqual('76623714-0')
+            expect('91411100766237140X'.usciToOibc).toEqual('76623714-0')
             expect(Convertor.usciToOibc('91411100766237140')).toEqual('')
         })
         it('日期转换', function () {
@@ -53,10 +59,15 @@ try {
             expect(Convertor.timeFormat(new Date(), 'yy年M月d日 h:m:s') == '').toEqual(false)
             expect(Convertor.timeFormat(new Date(), 'qq S') == '').toEqual(false)
             expect(Convertor.timeFormat(new Date()) == '').toEqual(false)
+            expect(new Date().timeFormat() == '').toEqual(false)
+            expect(new Date().timeFormat('yy年M月d日 h:m:s') == '').toEqual(false)
         })
         it('千分位处理', function () {
+            let num = 10009992.12
             expect(Convertor.thousands(10009992.12)).toEqual('10,009,992.12')
             expect(Convertor.thousands(1)).toEqual('1')
+            expect('10009992.12'.thousands).toEqual('10,009,992.12')
+            expect(num.thousands).toEqual('10,009,992.12')
         })
         it('文本转base64', function () {
             expect(Convertor.textToBase64('1234')).toEqual('MTIzNA==')
@@ -81,27 +92,33 @@ try {
             expect(Convertor.hexToRgb('#f2a')).toEqual('rgb(255,34,170)')
             expect(Convertor.hexToRgb('#ff22aa')).toEqual('rgb(255,34,170)')
             expect(Convertor.hexToRgb('ff22aa')).toEqual('rgb(255,34,170)')
+            expect('ff22aa'.hexToRgb).toEqual('rgb(255,34,170)')
             expect(() => Convertor.hexToRgb('')).toThrow();
         })
         it('颜色转换', function () {
             expect(Convertor.rgbToHex('rgb(255,34,170)')).toEqual('#ff22aa')
             expect(Convertor.rgbToHex('rgb(1,34,170)')).toEqual('#0122aa')
+            expect('rgb(255,34,170)'.rgbToHex).toEqual('#ff22aa')
             expect(() => Convertor.rgbToHex('')).toThrow();
         })
         it('xml输出文本', function () {
             expect(Convertor.xmlToText('<div>总金额 <span>100,000.00</span></div>')).toEqual('总金额 100,000.00')
+            expect('<div>总金额 <span>100,000.00</span></div>'.xmlToText).toEqual('总金额 100,000.00')
         })
         it('数字转大写人民币', function () {
             expect(Convertor.numToAmountInWords(102030.00)).toEqual('壹拾万贰仟零叁拾圆整')
+            expect('102030'.numToAmountInWords).toEqual('壹拾万贰仟零叁拾圆整')
             expect(Convertor.numToAmountInWords(-102030.00)).toEqual('欠壹拾万贰仟零叁拾圆整')
             expect(Convertor.numToAmountInWords(undefined)).toEqual('')
         })
         it('数字转中文', function () {
             expect(Convertor.numToChinese(102030)).toEqual('十万二千三十')
+            expect('102030'.numToChinese).toEqual('十万二千三十')
             expect(Convertor.numToChinese(undefined)).toEqual('')
         })
         it('url query 转对象', function () {
             expect(Convertor.urlQueryToObject('http://example.com/user?id=1&age=2')).toEqual({ "age": "2", "id": "1" })
+            expect('http://example.com/user?id=1&age=2'.urlQueryToObject).toEqual({ "age": "2", "id": "1" })
             expect(Convertor.urlQueryToObject('')).toEqual({})
         })
         it('url object 转 query', function () {
@@ -109,12 +126,15 @@ try {
         })
         it('蛇形命名法 -> 大驼峰命名法', function () {
             expect(Convertor.snakeCaseToUpperCamelcase('lower_case_with_underscores')).toEqual('LowerCaseWithUnderscores')
+            expect('lower_case_with_underscores'.snakeCaseToUpperCamelcase).toEqual('LowerCaseWithUnderscores')
         })
         it('蛇形命名法 -> 小驼峰命名法', function () {
             expect(Convertor.snakeCaseToLowerCamelcase('lower_case_with_underscores')).toEqual('lowerCaseWithUnderscores')
+            expect('lower_case_with_underscores'.snakeCaseToLowerCamelcase).toEqual('lowerCaseWithUnderscores')
         })
         it('驼峰命名法 -> 蛇形命名法', function () {
             expect(Convertor.camelcaseToSnakeCase('LowerCaseWithUnderscores')).toEqual('lower_case_with_underscores')
+            expect('LowerCaseWithUnderscores'.camelcaseToSnakeCase).toEqual('lower_case_with_underscores')
         })
         it('通过日期获取星座', function () {
             expect(Convertor.getConstellationByDate('01/01')).toEqual('摩羯座')
@@ -130,8 +150,33 @@ try {
             expect(Convertor.getConstellationByDate('11/01')).toEqual('天蝎座')
             expect(Convertor.getConstellationByDate('12/01')).toEqual('射手座')
             expect(Convertor.getConstellationByDate('12/24')).toEqual('摩羯座')
+            expect('12/24'.getConstellationByDate).toEqual('摩羯座')
             expect(Convertor.getConstellationByDate(new Date(2000, 11, 24))).toEqual('摩羯座')
+            expect(new Date(2000, 11, 24).getConstellationByDate).toEqual('摩羯座')
             expect(Convertor.getConstellationByDate(1224)).toEqual('摩羯座')
+        })
+        it('身份证号解析', function () {
+            expect(Convertor.citizenIdentificationNumberParse('230504199607116664')).toEqual({
+                reginCode: '230504',
+                birthday: '1996/07/11',
+                gender: '女',
+                constellation: '巨蟹座',
+                age: 27
+            })
+            expect('230504199607116664'.citizenIdentificationNumberParse).toEqual({
+                reginCode: '230504',
+                birthday: '1996/07/11',
+                gender: '女',
+                constellation: '巨蟹座',
+                age: 27
+            })
+            expect('420503198804097532'.citizenIdentificationNumberParse).toEqual({
+                "age": 35,
+                "birthday": "1988/04/09",
+                "constellation": "白羊座",
+                "gender": "男",
+                "reginCode": "420503",
+            })
         })
 
     })